advertising-management-at-gas-stations-develop/src/app/backend/requirements.txt
aiodns==3.2.0
aiofile==3.9.0
aiofiles==24.1.0
aiopath==0.7.7
aioping==0.4.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
asyncpg==0.30.0
asyncssh==2.19.0
bcrypt==4.2.1
caio==0.9.21
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
click-spinner==0.1.10
colorama==0.4.6
croniter==1.4.1
cryptography==44.0.0
decorator==5.1.1
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastapi-utilities==0.3.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moviepy==2.1.2
numpy==2.2.2
paramiko==3.5.0
passlib==1.7.4
pillow==10.4.0
proglog==0.1.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycares==4.5.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
python-telegram-bot==21.9
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.3
websockets==14.1
